% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar-iffhs.R
\name{baixar_iffhs}
\alias{baixar_iffhs}
\title{Baixa e limpa o dataset da IFFHS}
\usage{
baixar_iffhs(ano_inicio_carreira = 1900)
}
\arguments{
\item{ano_inicio_carreira}{Ano de início da carreira do jogador}
}
\value{
Um tibble com os dados da IFFHS
}
\description{
Esta função baixa o dataset da IFFHS e faz a limpeza dos dados
}
